fix(bench_saturate): fix GCC -Wconversion narrowing in sender round-robin
vector::size() returns size_t (64-bit); explicit cast to uint32_t
prevents -Werror=conversion failure on GCC.

+++ apps/bench_saturate/actors/saturate_sender_actor.hpp
@@ -178,7 +178,8 @@ class SaturateSenderActor : public EventBasedActor {
             }
 
             auto& target = receiver_addrs_[next_receiver_idx_];
-            next_receiver_idx_ = (next_receiver_idx_ + 1) % receiver_addrs_.size();
+            next_receiver_idx_ = static_cast<uint32_t>((next_receiver_idx_ + 1) %
+                                                       receiver_addrs_.size());
 
             context()->send(target, make_msg(LoadMessageTag, std::move(payload)));
             sent_count_.fetch_add(1, std::memory_order_relaxed);
